refactor(video): update import paths for video component props and create video.props.tsx
