+++ examples/kit-nextjs-article-starter/src/components/video/video.props.tsx
